refactor(evolution): isolate evaluation batch boundary
